remove unused queries
