+++ apps/desktop/src/hooks/queries/useBeats.ts
@@ -19,7 +19,6 @@ import {
     NewBeatArgs,
 } from "@/db-functions";
 import { DEFAULT_STALE_TIME } from "./constants";
-import { invalidateAllAppearances } from "./sharedInvalidators";
 
 /**
  * Arguments for shifting beats
@@ -78,8 +77,6 @@ export const createBeatsMutationOptions = (qc: QueryClient) => {
             void qc.invalidateQueries({
                 queryKey: [KEY_BASE],
             });
-            // New beats shift later page timestamps
-            invalidateAllAppearances(qc);
         },
         onError: (e, variables) => {
             conToastError(`Error creating beats`, e, variables);
@@ -91,16 +88,11 @@ export const updateBeatsMutationOptions = (qc: QueryClient) => {
     return mutationOptions({
         mutationFn: (modifiedBeats: ModifiedBeatArgs[]) =>
             updateBeats({ db, modifiedBeats }),
-        onSuccess: (_, variables) => {
+        onSuccess: () => {
             // Invalidate all beat queries
-            const beatIds = new Set<number>();
-            for (const modifiedArgs of variables) beatIds.add(modifiedArgs.id);
-
             void qc.invalidateQueries({
                 queryKey: [KEY_BASE],
             });
-            // A beat's duration/tempo shifts later page timestamps
-            invalidateAllAppearances(qc);
         },
         onError: (e, variables) => {
             conToastError(`Error updating beats`, e, variables);
@@ -116,8 +108,6 @@ export const deleteBeatsMutationOptions = (qc: QueryClient) => {
             void qc.invalidateQueries({
                 queryKey: [KEY_BASE],
             });
-            // Removing beats shifts later page timestamps
-            invalidateAllAppearances(qc);
         },
         onError: (e, variables) => {
             conToastError(`Error deleting beats`, e, variables);
@@ -133,8 +123,6 @@ export const shiftBeatsMutationOptions = (qc: QueryClient) => {
             void qc.invalidateQueries({
                 queryKey: [KEY_BASE],
             });
-            // Shifted beats shift page timestamps
-            invalidateAllAppearances(qc);
         },
         onError: (e, variables) => {
             conToastError(`Error shifting beats`, e, variables);
@@ -150,8 +138,6 @@ export const flattenOrderMutationOptions = (qc: QueryClient) => {
             void qc.invalidateQueries({
                 queryKey: [KEY_BASE],
             });
-            // Reordered beats shift page timestamps
-            invalidateAllAppearances(qc);
         },
         onError: (e) => {
             conToastError(`Error flattening beat order`, e);
