+++ apps/desktop/src/hooks/queries/useSectionAppearances.ts
@@ -19,7 +19,6 @@ import {
 } from "@/db-functions";
 import { DEFAULT_STALE_TIME } from "./constants";
 import { marcherWithVisualsKeys } from "./useMarchersWithVisuals";
-import { invalidateAllAppearances } from "./sharedInvalidators";
 
 const KEY_BASE = "section_appearances";
 
@@ -117,9 +116,6 @@ export const updateSectionAppearancesMutationOptions = (qc: QueryClient) => {
             await qc.invalidateQueries({
                 queryKey: sectionAppearanceKeys.all(),
             });
-            // A section's colors can affect any marcher in that section; the
-            // modified args don't reliably include which section, so invalidate broadly.
-            invalidateAllAppearances(qc);
         },
         onError: (e, variables) => {
             conToastError(`Error updating section appearances`, e, variables);
